build(scaffold): copier update v0.13.0 -> v0.16.0
Adopts the maximal guardrail superset (sdlc-scaffold v0.14 vip epic),
the enforced magic-literals ratchet (v0.15), and the advisory
main-sequence / shape-contract gates (v0.16).

New enforced gates handled:
- ruff union select (S101, N, RUF0xx, PLR, SLF001, PTH, PERF, …): 178
  hits -> 0. Resolutions:
  * N806/N803 (43): ML linalg + feature-matrix idioms folded into the
    pep8-naming ignore-names LOCAL-SLOT (X*/F*/C*/P*/R*/W*, single-cap
    A/M/T/…); genuine non-idioms renamed (logT->log_t, Net->net_cls,
    LATE->late_hits).
  * SLF001 (99): a scaffold-owned rule conflict — the mandated
    op-namespace pattern (ast-grep py-top-level-function) forces
    Cls._helper access on sibling statics + module-level registries,
    which SLF001 flags though it is no cross-boundary reach-in. Carved
    in the per-file-ignores LOCAL-SLOT (runner tree + static-op-namespace
    core files); filed sdlc-scaffold-8ex for a proper select/ignore slot.
  * RUF046/RUF022/C420 (auto), plus PERF401/PLW0108/RUF007/RUF005/PTH123
    hand-fixed. fnirs_windowed_eval _ARMS rebuilt as one comprehension
    assignment (satisfies PERF401 without tripping py-top-level-side-effect).
- magic-literals (enforced): ceiling frozen at mindscape's DataFrame-
  schema + metric-key floor (max_strings=33, max_key_sets=10) — the
  design's named non-enumerable floor; ratchets down.

Merge resolution: kept the ML dep tree + uv cu130 index (local mods) and
added jaxtyping/beartype (shape-contract runtime checker); folded vulture
ignore_names into the new vulture-ignore-names slot; CLAUDE/AGENTS/PLAN
--ours then synced the ruff + new-gate docs.

Gates green: nox lint + test + cov (301 passed, 92% >= 80 floor).

+++ core/data/eeg/braindecode_pre.py
@@ -63,7 +63,7 @@ class BraindecodePre:
                                       init_block_size=cfg.init_block_size))
         preprocess(ds, steps)
 
-        start = int(round(cfg.trial_start_offset_s * sfreq))
+        start = round(cfg.trial_start_offset_s * sfreq)
         windows = create_windows_from_events(ds, trial_start_offset_samples=start,
                                              trial_stop_offset_samples=0, preload=True)
